refactor(all): cleanup & optimize for size
