+++ packages/kernel/src/di.ts
@@ -3,15 +3,16 @@
 /* eslint-disable @typescript-eslint/strict-boolean-expressions */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { isObject, applyMetadataPolyfill } from '@aurelia/metadata';
+import { applyMetadataPolyfill } from '@aurelia/metadata';
 
 applyMetadataPolyfill(Reflect, false, false);
 
-import { isArrayIndex, isNativeFunction } from './functions';
-import { Class, Constructable, IDisposable } from './interfaces';
-import { emptyArray } from './platform';
-import { appendAnnotation, getAnnotationKeyFor, IResourceKind, Protocol, ResourceDefinition, ResourceType } from './resource';
-import { createError, createObject, defineMetadata, getOwnMetadata, isFunction, isString, toStringSafe } from './utilities';
+import { isArrayIndex } from './functions';
+import { Container } from './di.container';
+import { Constructable, IDisposable, Writable } from './interfaces';
+import { appendAnnotation, getAnnotationKeyFor, IResourceKind, ResourceDefinition, ResourceType } from './resource';
+import { createError, defineMetadata, getOwnMetadata, isFunction, isString, toStringSafe } from './utilities';
+import { instanceRegistration, singletonRegistration, transientRegistation, callbackRegistration, cachedCallbackRegistration, aliasToRegistration, deferRegistration, cacheCallbackResult } from './di.registration';
 
 export type ResolveCallback<T = any> = (handler: IContainer, requestor: IContainer, resolver: IResolver<T>) => T;
 
@@ -136,7 +137,7 @@ export type Resolved<K> = (
 
 export type Injectable<T = {}> = Constructable<T> & { inject?: Key[] };
 
-function cloneArrayWithPossibleProps<T>(source: readonly T[]): T[] {
+const cloneArrayWithPossibleProps = <T>(source: readonly T[]): T[] => {
   const clone = source.slice();
   const keys = Object.keys(source);
   const len = keys.length;
@@ -148,7 +149,7 @@ function cloneArrayWithPossibleProps<T>(source: readonly T[]): T[] {
     }
   }
   return clone;
-}
+};
 
 export interface IContainerConfiguration {
   /**
@@ -166,8 +167,8 @@ export const DefaultResolver = {
   none(key: Key): IResolver {
     throw noResolverForKeyError(key);
   },
-  singleton(key: Key): IResolver {return new Resolver(key, ResolverStrategy.singleton, key);},
-  transient(key: Key): IResolver {return new Resolver(key, ResolverStrategy.transient, key);},
+  singleton: (key: Key): IResolver => new Resolver(key, ResolverStrategy.singleton, key),
+  transient: (key: Key): IResolver => new Resolver(key, ResolverStrategy.transient, key),
 };
 
 const noResolverForKeyError = (key: Key) =>
@@ -197,18 +198,138 @@ export class ContainerConfiguration implements IContainerConfiguration {
   }
 }
 
+/** @internal */
+export const createContainer = (config?: Partial<IContainerConfiguration>): IContainer => new Container(null, ContainerConfiguration.from(config));
+
+const getAnnotationParamtypes = (Type: Constructable | Injectable): readonly Key[] | undefined => {
+  const key = getAnnotationKeyFor('di:paramtypes');
+  return getOwnMetadata(key, Type);
+};
+
+const getDesignParamtypes = (Type: Constructable | Injectable): readonly Key[] | undefined =>
+  getOwnMetadata('design:paramtypes', Type);
+
+const getOrCreateAnnotationParamTypes = (Type: Constructable | Injectable): Key[] => {
+  const key = getAnnotationKeyFor('di:paramtypes');
+  let annotationParamtypes = getOwnMetadata(key, Type);
+  if (annotationParamtypes === void 0) {
+    defineMetadata(key, annotationParamtypes = [], Type);
+    appendAnnotation(Type, key);
+  }
+  return annotationParamtypes;
+};
+
+export const getDependencies = (Type: Constructable | Injectable): Key[] => {
+  // Note: Every detail of this getDependencies method is pretty deliberate at the moment, and probably not yet 100% tested from every possible angle,
+  // so be careful with making changes here as it can have a huge impact on complex end user apps.
+  // Preferably, only make changes to the dependency resolution process via a RFC.
+
+  const key = getAnnotationKeyFor('di:dependencies');
+  let dependencies = getOwnMetadata(key, Type) as Key[] | undefined;
+  if (dependencies === void 0) {
+    // Type.length is the number of constructor parameters. If this is 0, it could mean the class has an empty constructor
+    // but it could also mean the class has no constructor at all (in which case it inherits the constructor from the prototype).
+
+    // Non-zero constructor length + no paramtypes means emitDecoratorMetadata is off, or the class has no decorator.
+    // We're not doing anything with the above right now, but it's good to keep in mind for any future issues.
+
+    const inject = (Type as Injectable).inject;
+    if (inject === void 0) {
+      // design:paramtypes is set by tsc when emitDecoratorMetadata is enabled.
+      const designParamtypes = getDesignParamtypes(Type);
+      // au:annotation:di:paramtypes is set by the parameter decorator from DI.createInterface or by @inject
+      const annotationParamtypes = getAnnotationParamtypes(Type);
+      if (designParamtypes === void 0) {
+        if (annotationParamtypes === void 0) {
+          // Only go up the prototype if neither static inject nor any of the paramtypes is defined, as
+          // there is no sound way to merge a type's deps with its prototype's deps
+          const Proto = Object.getPrototypeOf(Type);
+          if (isFunction(Proto) && Proto !== Function.prototype) {
+            dependencies = cloneArrayWithPossibleProps(getDependencies(Proto));
+          } else {
+            dependencies = [];
+          }
+        } else {
+          // No design:paramtypes so just use the au:annotation:di:paramtypes
+          dependencies = cloneArrayWithPossibleProps(annotationParamtypes);
+        }
+      } else if (annotationParamtypes === void 0) {
+        // No au:annotation:di:paramtypes so just use the design:paramtypes
+        dependencies = cloneArrayWithPossibleProps(designParamtypes);
+      } else {
+        // We've got both, so merge them (in case of conflict on same index, au:annotation:di:paramtypes take precedence)
+        dependencies = cloneArrayWithPossibleProps(designParamtypes);
+        let len = annotationParamtypes.length;
+        let auAnnotationParamtype: Key;
+        let i = 0;
+        for (; i < len; ++i) {
+          auAnnotationParamtype = annotationParamtypes[i];
+          if (auAnnotationParamtype !== void 0) {
+            dependencies[i] = auAnnotationParamtype;
+          }
+        }
+
+        const keys = Object.keys(annotationParamtypes);
+        let key: string;
+        i = 0;
+        len = keys.length;
+        for (i = 0; i < len; ++i) {
+          key = keys[i];
+          if (!isArrayIndex(key)) {
+            dependencies[key] = annotationParamtypes[key];
+          }
+        }
+      }
+    } else {
+      // Ignore paramtypes if we have static inject
+      dependencies = cloneArrayWithPossibleProps(inject);
+    }
+
+    defineMetadata(key, dependencies, Type);
+    appendAnnotation(Type, key);
+  }
+
+  return dependencies;
+};
+
+/**
+ * @internal
+ *
+ * @param configureOrName - Use for improving error messaging
+ */
+export const createInterface = <K extends Key>(configureOrName?: string | ((builder: ResolverBuilder<K>) => IResolver<K>), configuror?: (builder: ResolverBuilder<K>) => IResolver<K>): InterfaceSymbol<K> => {
+ const configure = isFunction(configureOrName) ? configureOrName : configuror;
+ const friendlyName = isString(configureOrName) ? configureOrName : undefined;
+
+ const Interface = function (target: Injectable<K>, property: string, index: number): void {
+   if (target == null || new.target !== undefined) {
+    throw createNoRegistrationError(Interface.friendlyName);
+   }
+   const annotationParamtypes = getOrCreateAnnotationParamTypes(target);
+   annotationParamtypes[index] = Interface;
+ };
+ Interface.$isInterface = true;
+ Interface.friendlyName = friendlyName == null ? '(anonymous)' : friendlyName;
+
+ if (configure != null) {
+   Interface.register = (container: IContainer, key?: Key): IResolver<K> =>
+     configure(new ResolverBuilder(container, key ?? Interface));
+ }
+
+ Interface.toString = (): string => `InterfaceSymbol<${Interface.friendlyName}>`;
+
+ return Interface;
+};
+const createNoRegistrationError = (name: string) =>
+  __DEV__
+    ? createError(`AUR0001: No registration for interface: '${name}'`)
+    : createError(`AUR0001:${name}`);
+
 export const DI = {
-  createContainer(config?: Partial<IContainerConfiguration>): IContainer {
-    return new Container(null, ContainerConfiguration.from(config));
-  },
-  getDesignParamtypes(Type: Constructable | Injectable): readonly Key[] | undefined {
-    return getOwnMetadata('design:paramtypes', Type);
-  },
-  getAnnotationParamtypes(Type: Constructable | Injectable): readonly Key[] | undefined {
-    const key = getAnnotationKeyFor('di:paramtypes');
-    return getOwnMetadata(key, Type);
-  },
-  getOrCreateAnnotationParamTypes: getOrCreateAnnotationParamTypes,
+  createContainer,
+  getDesignParamtypes,
+  getAnnotationParamtypes,
+  getOrCreateAnnotationParamTypes,
   getDependencies: getDependencies,
   /**
    * creates a decorator that also matches an interface and can be used as a {@linkcode Key}.
@@ -249,37 +370,11 @@ export const DI = {
    * container.get(Foo).str; // returns 'somestring'
    * ```
    *
-   * - @param friendlyName used to improve error messaging
+   * - @param configureOrName - supply a string to improve error messaging
    */
-  createInterface<K extends Key>(configureOrName?: string | ((builder: ResolverBuilder<K>) => IResolver<K>), configuror?: (builder: ResolverBuilder<K>) => IResolver<K>): InterfaceSymbol<K> {
-    const configure = isFunction(configureOrName) ? configureOrName : configuror;
-    const friendlyName = isString(configureOrName) ? configureOrName : undefined;
-
-    const Interface = function (target: Injectable<K>, property: string, index: number): void {
-      if (target == null || new.target !== undefined) {
-        if (__DEV__) {
-          throw createError(`AUR0001: No registration for interface: '${Interface.friendlyName}'`);
-        } else {
-          throw createError(`AUR0001:${Interface.friendlyName}`);
-        }
-      }
-      const annotationParamtypes = getOrCreateAnnotationParamTypes(target);
-      annotationParamtypes[index] = Interface;
-    };
-    Interface.$isInterface = true;
-    Interface.friendlyName = friendlyName == null ? '(anonymous)' : friendlyName;
-
-    if (configure != null) {
-      Interface.register = (container: IContainer, key?: Key): IResolver<K> =>
-        configure(new ResolverBuilder(container, key ?? Interface));
-    }
-
-    Interface.toString = (): string => `InterfaceSymbol<${Interface.friendlyName}>`;
-
-    return Interface;
-  },
+  createInterface,
   inject(...dependencies: Key[]): (target: Injectable, key?: string | number, descriptor?: PropertyDescriptor | number) => void {
-    return function (target: Injectable, key?: string | number, descriptor?: PropertyDescriptor | number): void {
+    return (target: Injectable, key?: string | number, descriptor?: PropertyDescriptor | number): void => {
       if (typeof descriptor === 'number') { // It's a parameter decorator.
         const annotationParamtypes = getOrCreateAnnotationParamTypes(target);
         const dep = dependencies[0];
@@ -370,96 +465,13 @@ export const DI = {
   },
 };
 
-function getDependencies(Type: Constructable | Injectable): Key[] {
-  // Note: Every detail of this getDependencies method is pretty deliberate at the moment, and probably not yet 100% tested from every possible angle,
-  // so be careful with making changes here as it can have a huge impact on complex end user apps.
-  // Preferably, only make changes to the dependency resolution process via a RFC.
-
-  const key = getAnnotationKeyFor('di:dependencies');
-  let dependencies = getOwnMetadata(key, Type) as Key[] | undefined;
-  if (dependencies === void 0) {
-    // Type.length is the number of constructor parameters. If this is 0, it could mean the class has an empty constructor
-    // but it could also mean the class has no constructor at all (in which case it inherits the constructor from the prototype).
-
-    // Non-zero constructor length + no paramtypes means emitDecoratorMetadata is off, or the class has no decorator.
-    // We're not doing anything with the above right now, but it's good to keep in mind for any future issues.
-
-    const inject = (Type as Injectable).inject;
-    if (inject === void 0) {
-      // design:paramtypes is set by tsc when emitDecoratorMetadata is enabled.
-      const designParamtypes = DI.getDesignParamtypes(Type);
-      // au:annotation:di:paramtypes is set by the parameter decorator from DI.createInterface or by @inject
-      const annotationParamtypes = DI.getAnnotationParamtypes(Type);
-      if (designParamtypes === void 0) {
-        if (annotationParamtypes === void 0) {
-          // Only go up the prototype if neither static inject nor any of the paramtypes is defined, as
-          // there is no sound way to merge a type's deps with its prototype's deps
-          const Proto = Object.getPrototypeOf(Type);
-          if (isFunction(Proto) && Proto !== Function.prototype) {
-            dependencies = cloneArrayWithPossibleProps(getDependencies(Proto));
-          } else {
-            dependencies = [];
-          }
-        } else {
-          // No design:paramtypes so just use the au:annotation:di:paramtypes
-          dependencies = cloneArrayWithPossibleProps(annotationParamtypes);
-        }
-      } else if (annotationParamtypes === void 0) {
-        // No au:annotation:di:paramtypes so just use the design:paramtypes
-        dependencies = cloneArrayWithPossibleProps(designParamtypes);
-      } else {
-        // We've got both, so merge them (in case of conflict on same index, au:annotation:di:paramtypes take precedence)
-        dependencies = cloneArrayWithPossibleProps(designParamtypes);
-        let len = annotationParamtypes.length;
-        let auAnnotationParamtype: Key;
-        let i = 0;
-        for (; i < len; ++i) {
-          auAnnotationParamtype = annotationParamtypes[i];
-          if (auAnnotationParamtype !== void 0) {
-            dependencies[i] = auAnnotationParamtype;
-          }
-        }
-
-        const keys = Object.keys(annotationParamtypes);
-        let key: string;
-        i = 0;
-        len = keys.length;
-        for (i = 0; i < len; ++i) {
-          key = keys[i];
-          if (!isArrayIndex(key)) {
-            dependencies[key] = annotationParamtypes[key];
-          }
-        }
-      }
-    } else {
-      // Ignore paramtypes if we have static inject
-      dependencies = cloneArrayWithPossibleProps(inject);
-    }
-
-    defineMetadata(key, dependencies, Type);
-    appendAnnotation(Type, key);
-  }
-
-  return dependencies;
-}
-
-function getOrCreateAnnotationParamTypes(Type: Constructable | Injectable): Key[] {
-  const key = getAnnotationKeyFor('di:paramtypes');
-  let annotationParamtypes = getOwnMetadata(key, Type);
-  if (annotationParamtypes === void 0) {
-    defineMetadata(key, annotationParamtypes = [], Type);
-    appendAnnotation(Type, key);
-  }
-  return annotationParamtypes;
-}
-
-export const IContainer = DI.createInterface<IContainer>('IContainer');
+export const IContainer = createInterface<IContainer>('IContainer');
 export const IServiceLocator = IContainer as unknown as InterfaceSymbol<IServiceLocator>;
 
 function createResolver(getter: (key: any, handler: IContainer, requestor: IContainer) => any): (key: any) => any {
   return function (key: any): ReturnType<typeof DI.inject> {
     const resolver: ReturnType<typeof DI.inject> & Partial<Pick<IResolver, 'resolve'>> & { $isResolver: true} = function (target: Injectable, property?: string | number, descriptor?: PropertyDescriptor | number): void {
-      DI.inject(resolver)(target, property, descriptor);
+      inject(resolver)(target, property, descriptor);
     };
 
     resolver.$isResolver = true;
@@ -505,10 +517,11 @@ export function transient<T extends Constructable>(target?: T & Partial<Register
 
 type SingletonOptions = { scoped: boolean };
 const defaultSingletonOptions = { scoped: false };
+const decorateSingleton = DI.singleton;
 
-function singletonDecorator<T extends Constructable>(target: T & Partial<RegisterSelf<T>>): T & RegisterSelf<T> {
-  return DI.singleton(target);
-}
+const singletonDecorator = <T extends Constructable>(target: T & Partial<RegisterSelf<T>>): T & RegisterSelf<T> => {
+  return decorateSingleton(target);
+};
 /**
  * Registers the decorated class as a singleton dependency; the class will only be created once. Each
  * consecutive time the dependency is resolved, the same instance will be returned.
@@ -534,17 +547,17 @@ export function singleton<T extends Constructable>(options?: SingletonOptions): 
 export function singleton<T extends Constructable>(target: T & Partial<RegisterSelf<T>>): T & RegisterSelf<T>;
 export function singleton<T extends Constructable>(targetOrOptions?: (T & Partial<RegisterSelf<T>>) | SingletonOptions): T & RegisterSelf<T> | typeof singletonDecorator {
   if (isFunction(targetOrOptions)) {
-    return DI.singleton(targetOrOptions);
+    return decorateSingleton(targetOrOptions);
   }
   return function <T extends Constructable>($target: T) {
-    return DI.singleton($target, targetOrOptions);
+    return decorateSingleton($target, targetOrOptions);
   };
 }
 
-function createAllResolver(
+const createAllResolver = (
   getter: (key: any, handler: IContainer, requestor: IContainer, searchAncestors: boolean) => readonly any[]
-): (key: any, searchAncestors?: boolean) => ReturnType<typeof DI.inject> {
-  return function (key: any, searchAncestors?: boolean): ReturnType<typeof DI.inject> {
+): (key: any, searchAncestors?: boolean) => ReturnType<typeof DI.inject> => {
+  return (key: any, searchAncestors?: boolean): ReturnType<typeof DI.inject> => {
     searchAncestors = !!searchAncestors;
     const resolver: ReturnType<typeof DI.inject>
       & Required<Pick<IResolver, 'resolve'>>
@@ -554,7 +567,7 @@ function createAllResolver(
           property?: string | number,
           descriptor?: PropertyDescriptor | number
         ): void {
-          DI.inject(resolver)(target, property, descriptor);
+          inject(resolver)(target, property, descriptor);
         };
 
     resolver.$isResolver = true;
@@ -564,7 +577,7 @@ function createAllResolver(
 
     return resolver;
   };
-}
+};
 
 export const all = createAllResolver(
   (key: any,
@@ -643,9 +656,9 @@ export const optional = createResolver((key: Key, handler: IContainer, requestor
 /**
  * ignore tells the container not to try to inject a dependency
  */
-export function ignore(target: Injectable, property?: string | number, descriptor?: PropertyDescriptor | number): void {
-  DI.inject(ignore)(target, property, descriptor);
-}
+export const ignore = (target: Injectable, property?: string | number, descriptor?: PropertyDescriptor | number): void => {
+  inject(ignore)(target, property, descriptor);
+};
 ignore.$isResolver = true;
 ignore.resolve = () => undefined;
 
@@ -711,12 +724,12 @@ export type INewInstanceResolver<T> = {
   (...args: unknown[]): any;
 };
 
-function createNewInstance(key: any, handler: IContainer, requestor: IContainer) {
+const createNewInstance = (key: any, handler: IContainer, requestor: IContainer) => {
   return handler.getFactory(key).construct(requestor);
-}
+};
 
 _START_CONST_ENUM();
-const enum ResolverStrategy {
+export const enum ResolverStrategy {
   instance = 0,
   singleton = 1,
   transient = 2,
@@ -726,7 +739,7 @@ const enum ResolverStrategy {
 }
 _END_CONST_ENUM();
 
-class Resolver implements IResolver, IRegistration {
+export class Resolver implements IResolver, IRegistration {
   public constructor(
     public _key: Key,
     public _strategy: ResolverStrategy,
@@ -810,7 +823,7 @@ export interface IInvoker<T extends Constructable = any> {
   ): Resolved<T>;
 }
 
-function containerGetKey(this: IContainer, d: Key) {
+export function containerGetKey(this: IContainer, d: Key) {
   return this.get(d);
 }
 
@@ -846,533 +859,6 @@ export class Factory<T extends Constructable = any> implements IFactory<T> {
   }
 }
 
-const containerResolver: IResolver = {
-  $isResolver: true,
-  resolve(handler: IContainer, requestor: IContainer): IContainer {
-    return requestor;
-  }
-};
-
-function isRegistry(obj: IRegistry | Record<string, IRegistry>): obj is IRegistry {
-  return isFunction(obj.register);
-}
-
-function isSelfRegistry<T extends Constructable>(obj: RegisterSelf<T>): obj is RegisterSelf<T> {
-  return isRegistry(obj) && typeof obj.registerInRequestor === 'boolean';
-}
-
-function isRegisterInRequester<T extends Constructable>(obj: RegisterSelf<T>): obj is RegisterSelf<T> {
-  return isSelfRegistry(obj) && obj.registerInRequestor;
-}
-
-function isClass<T extends { prototype?: any }>(obj: T): obj is Class<any, T> {
-  return obj.prototype !== void 0;
-}
-
-function isResourceKey(key: Key): key is string {
-  return isString(key) && key.indexOf(':') > 0;
-}
-
-const InstrinsicTypeNames = new Set<string>('Array ArrayBuffer Boolean DataView Date Error EvalError Float32Array Float64Array Function Int8Array Int16Array Int32Array Map Number Object Promise RangeError ReferenceError RegExp Set SharedArrayBuffer String SyntaxError TypeError Uint8Array Uint8ClampedArray Uint16Array Uint32Array URIError WeakMap WeakSet'.split(' '));
-
-// const factoryKey = 'di:factory';
-// const factoryAnnotationKey = Protocol.annotation.keyFor(factoryKey);
-let containerId = 0;
-/** @internal */
-export class Container implements IContainer {
-  public readonly id: number = ++containerId;
-  /** @internal */
-  private _registerDepth: number = 0;
-
-  public get depth(): number {
-    return this.parent === null ? 0 : this.parent.depth + 1;
-  }
-  public readonly root: Container;
-
-  /**
-   * All own resolvers of this container
-   *
-   * @internal
-   */
-  private readonly _resolvers: Map<Key, IResolver | IDisposableResolver>;
-  /**
-   * A map of Factory per Constructor (Type) of this container tree.
-   *
-   * Factories are "global" per container tree
-   *
-   * @internal
-   */
-  private readonly _factories: Map<Constructable, Factory>;
-
-  /**
-   * A map of all resources resolver by their key
-   */
-  private res: Record<string, IResolver | IDisposableResolver | undefined>;
-
-  /** @internal */
-  private readonly _disposableResolvers: Map<Key, IDisposableResolver> = new Map<Key, IDisposableResolver>();
-
-  public constructor(
-    private readonly parent: Container | null,
-    private readonly config: ContainerConfiguration,
-  ) {
-    if (parent === null) {
-      this.root = this;
-
-      this._resolvers = new Map();
-      this._factories = new Map<Constructable, Factory>();
-
-      this.res = createObject();
-    } else {
-      this.root = parent.root;
-
-      this._resolvers = new Map();
-      this._factories = parent._factories;
-
-      if (config.inheritParentResources) {
-        this.res = Object.assign(
-          createObject(),
-          parent.res,
-          this.root.res,
-        );
-      } else {
-        this.res = createObject();
-      }
-    }
-
-    this._resolvers.set(IContainer, containerResolver);
-  }
-
-  public register(...params: any[]): IContainer {
-    if (++this._registerDepth === 100) {
-      throw registrationError(params);
-    }
-    let current: IRegistry | Record<string, IRegistry>;
-    let keys: string[];
-    let value: IRegistry;
-    let j: number;
-    let jj: number;
-    let i = 0;
-    // eslint-disable-next-line
-    let ii = params.length;
-    for (; i < ii; ++i) {
-      current = params[i];
-      if (!isObject(current)) {
-        continue;
-      }
-      if (isRegistry(current)) {
-        current.register(this);
-      } else if (Protocol.resource.has(current)) {
-        const defs = Protocol.resource.getAll(current);
-        if (defs.length === 1) {
-          // Fast path for the very common case
-          defs[0].register(this);
-        } else {
-          j = 0;
-          jj = defs.length;
-          while (jj > j) {
-            defs[j].register(this);
-            ++j;
-          }
-        }
-      } else if (isClass(current)) {
-        Registration.singleton(current, current as Constructable).register(this);
-      } else {
-        keys = Object.keys(current);
-        j = 0;
-        jj = keys.length;
-        for (; j < jj; ++j) {
-          value = current[keys[j]];
-          if (!isObject(value)) {
-            continue;
-          }
-          // note: we could remove this if-branch and call this.register directly
-          // - the extra check is just a perf tweak to create fewer unnecessary arrays by the spread operator
-          if (isRegistry(value)) {
-            value.register(this);
-          } else {
-            this.register(value);
-          }
-        }
-      }
-    }
-    --this._registerDepth;
-    return this;
-  }
-
-  public registerResolver<K extends Key, T = K>(key: K, resolver: IResolver<T>, isDisposable: boolean = false): IResolver<T> {
-    validateKey(key);
-
-    const resolvers = this._resolvers;
-    const result = resolvers.get(key);
-
-    if (result == null) {
-      resolvers.set(key, resolver);
-      if (isResourceKey(key)) {
-        if (this.res[key] !== void 0) {
-          throw resourceExistError(key);
-        }
-        this.res[key] = resolver;
-      }
-    } else if (result instanceof Resolver && result._strategy === ResolverStrategy.array) {
-      (result._state as IResolver[]).push(resolver);
-    } else {
-      resolvers.set(key, new Resolver(key, ResolverStrategy.array, [result, resolver]));
-    }
-
-    if (isDisposable) {
-      this._disposableResolvers.set(key, resolver as IDisposableResolver<T>);
-    }
-
-    return resolver;
-  }
-
-  // public deregisterResolverFor<K extends Key>(key: K, searchAncestors: boolean): void {
-  //   validateKey(key);
-
-  //   // eslint-disable-next-line @typescript-eslint/no-this-alias
-  //   let current: Container | null = this;
-  //   let resolver: IResolver | undefined;
-
-  //   while (current != null) {
-  //     resolver = current._resolvers.get(key);
-
-  //     if (resolver != null) {
-  //       current._resolvers.delete(key);
-  //       break;
-  //     }
-  //     if (current.parent == null) { return; }
-  //     current = searchAncestors ? current.parent : null;
-  //   }
-
-  //   if (resolver == null) { return; }
-  //   if (resolver instanceof Resolver && resolver.strategy === ResolverStrategy.array) {
-  //     throw createError('Cannot deregister a resolver with array strategy');
-  //   }
-  //   if (this._disposableResolvers.has(resolver as IDisposableResolver<K>)) {
-  //     (resolver as IDisposableResolver<K>).dispose();
-  //   }
-  //   if (isResourceKey(key)) {
-  //     // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-  //     delete this.res[key];
-  //   }
-  // }
-
-  public registerTransformer<K extends Key, T = K>(key: K, transformer: Transformer<T>): boolean {
-    const resolver = this.getResolver(key);
-
-    if (resolver == null) {
-      return false;
-    }
-
-    if (resolver.getFactory) {
-      const factory = resolver.getFactory(this);
-
-      if (factory == null) {
-        return false;
-      }
-
-      // This type cast is a bit of a hacky one, necessary due to the duplicity of IResolverLike.
-      // Problem is that that interface's type arg can be of type Key, but the getFactory method only works on
-      // type Constructable. So the return type of that optional method has this additional constraint, which
-      // seems to confuse the type checker.
-      factory.registerTransformer(transformer as unknown as Transformer<Constructable>);
-      return true;
-    }
-
-    return false;
-  }
-
-  public getResolver<K extends Key, T = K>(key: K | Key, autoRegister: boolean = true): IResolver<T> | null {
-    validateKey(key);
-
-    if ((key as unknown as IResolver).resolve !== void 0) {
-      return key as unknown as IResolver;
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    let current: Container = this;
-    let resolver: IResolver | undefined;
-    let handler: Container;
-
-    while (current != null) {
-      resolver = current._resolvers.get(key);
-
-      if (resolver == null) {
-        if (current.parent == null) {
-          handler = (isRegisterInRequester(key as unknown as RegisterSelf<Constructable>)) ? this : current;
-          return autoRegister ? this._jitRegister(key, handler) : null;
-        }
-
-        current = current.parent;
-      } else {
-        return resolver;
-      }
-    }
-
-    return null;
-  }
-
-  public has<K extends Key>(key: K, searchAncestors: boolean = false): boolean {
-    return this._resolvers.has(key)
-      ? true
-      : searchAncestors && this.parent != null
-        ? this.parent.has(key, true)
-        : false;
-  }
-
-  public get<K extends Key>(key: K): Resolved<K> {
-    validateKey(key);
-
-    if ((key as IResolver).$isResolver) {
-      return (key as IResolver).resolve(this, this);
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    let current: Container = this;
-    let resolver: IResolver | undefined;
-    let handler: Container;
-
-    while (current != null) {
-      resolver = current._resolvers.get(key);
-
-      if (resolver == null) {
-        if (current.parent == null) {
-          handler = (isRegisterInRequester(key as unknown as RegisterSelf<Constructable>)) ? this : current;
-          resolver = this._jitRegister(key, handler);
-          return resolver.resolve(current, this);
-        }
-
-        current = current.parent;
-      } else {
-        return resolver.resolve(current, this);
-      }
-    }
-
-    throw cantResolveKeyError(key);
-  }
-
-  public getAll<K extends Key>(key: K, searchAncestors: boolean = false): readonly Resolved<K>[] {
-    validateKey(key);
-
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    const requestor = this;
-    let current: Container | null = requestor;
-    let resolver: IResolver | undefined;
-
-    if (searchAncestors) {
-      let resolutions: any[] = emptyArray;
-      while (current != null) {
-        resolver = current._resolvers.get(key);
-        if (resolver != null) {
-          resolutions = resolutions.concat(buildAllResponse(resolver, current, requestor));
-        }
-        current = current.parent;
-      }
-      return resolutions;
-    } else {
-      while (current != null) {
-        resolver = current._resolvers.get(key);
-
-        if (resolver == null) {
-          current = current.parent;
-
-          if (current == null) {
-            return emptyArray;
-          }
-        } else {
-          return buildAllResponse(resolver, current, requestor);
-        }
-      }
-    }
-
-    return emptyArray;
-  }
-
-  public invoke<T, TDeps extends unknown[] = unknown[]>(Type: Constructable<T>, dynamicDependencies?: TDeps): T {
-    if (isNativeFunction(Type)) {
-      throw createNativeInvocationError(Type);
-    }
-    if (dynamicDependencies === void 0) {
-      return new Type(...getDependencies(Type).map(containerGetKey, this));
-    } else {
-      return new Type(...getDependencies(Type).map(containerGetKey, this), ...dynamicDependencies);
-    }
-  }
-
-  public getFactory<K extends Constructable>(Type: K): IFactory<K> {
-    let factory = this._factories.get(Type);
-    if (factory === void 0) {
-      if (isNativeFunction(Type)) {
-        throw createNativeInvocationError(Type);
-      }
-      this._factories.set(Type, factory = new Factory<K>(Type, getDependencies(Type)));
-    }
-    return factory;
-  }
-
-  public registerFactory<K extends Constructable>(key: K, factory: IFactory<K>): void {
-    this._factories.set(key, factory as Factory);
-  }
-
-  public createChild(config?: Partial<IContainerConfiguration>): IContainer {
-    if (config === void 0 && this.config.inheritParentResources) {
-      if (this.config === ContainerConfiguration.DEFAULT) {
-        return new Container(this, this.config);
-      }
-      return new Container(
-        this,
-        ContainerConfiguration.from({
-          ...this.config,
-          inheritParentResources: false,
-        }),
-      );
-    }
-    return new Container(this, ContainerConfiguration.from(config ?? this.config));
-  }
-
-  public disposeResolvers(): void {
-    const resolvers = this._resolvers;
-    const disposableResolvers = this._disposableResolvers;
-
-    let disposable: IDisposable;
-    let key: Key;
-
-    for ([key, disposable] of disposableResolvers.entries()) {
-      disposable.dispose();
-      resolvers.delete(key);
-    }
-    disposableResolvers.clear();
-  }
-
-  public find<TType extends ResourceType, TDef extends ResourceDefinition>(kind: IResourceKind<TType, TDef>, name: string): TDef | null {
-    const key = kind.keyFrom(name);
-    let resolver = this.res[key];
-    if (resolver === void 0) {
-      resolver = this.root.res[key];
-      if (resolver === void 0) {
-        return null;
-      }
-    }
-
-    if (resolver === null) {
-      return null;
-    }
-
-    if (isFunction(resolver.getFactory)) {
-      const factory = resolver.getFactory(this);
-      if (factory === null || factory === void 0) {
-        return null;
-      }
-
-      const definition = getOwnMetadata(kind.name, factory.Type);
-      if (definition === void 0) {
-        // TODO: we may want to log a warning here, or even throw. This would happen if a dependency is registered with a resource-like key
-        // but does not actually have a definition associated via the type's metadata. That *should* generally not happen.
-        return null;
-      }
-
-      return definition;
-    }
-
-    return null;
-  }
-
-  public create<TType extends ResourceType, TDef extends ResourceDefinition>(kind: IResourceKind<TType, TDef>, name: string): InstanceType<TType> | null {
-    const key = kind.keyFrom(name);
-    let resolver = this.res[key];
-    if (resolver === void 0) {
-      resolver = this.root.res[key];
-      if (resolver === void 0) {
-        return null;
-      }
-      return resolver.resolve(this.root, this) ?? null;
-    }
-    return resolver.resolve(this, this) ?? null;
-  }
-
-  public dispose(): void {
-    if (this._disposableResolvers.size > 0) {
-      this.disposeResolvers();
-    }
-    this._resolvers.clear();
-  }
-
-  /** @internal */
-  private _jitRegister(keyAsValue: any, handler: Container): IResolver {
-    if (!isFunction(keyAsValue)) {
-      throw jitRegisterNonFunctionError(keyAsValue);
-    }
-    if (InstrinsicTypeNames.has(keyAsValue.name)) {
-      throw jitInstrinsicTypeError(keyAsValue);
-    }
-
-    if (isRegistry(keyAsValue)) {
-      const registrationResolver = keyAsValue.register(handler, keyAsValue);
-      if (!(registrationResolver instanceof Object) || (registrationResolver as IResolver).resolve == null) {
-        const newResolver = handler._resolvers.get(keyAsValue);
-        if (newResolver != null) {
-          return newResolver;
-        }
-        throw invalidResolverFromRegisterError();
-      }
-      return registrationResolver as IResolver;
-    } else if (Protocol.resource.has(keyAsValue)) {
-      const defs = Protocol.resource.getAll(keyAsValue);
-      if (defs.length === 1) {
-        // Fast path for the very common case
-        defs[0].register(handler);
-      } else {
-        const len = defs.length;
-        for (let d = 0; d < len; ++d) {
-          defs[d].register(handler);
-        }
-      }
-      const newResolver = handler._resolvers.get(keyAsValue);
-      if (newResolver != null) {
-        return newResolver;
-      }
-      throw invalidResolverFromRegisterError();
-    } else if (keyAsValue.$isInterface) {
-      throw jitInterfaceError(keyAsValue.friendlyName);
-    } else {
-      const resolver = this.config.defaultResolver(keyAsValue, handler);
-      handler._resolvers.set(keyAsValue, resolver);
-      return resolver;
-    }
-  }
-}
-const registrationError = (deps: Key[]) =>
-  // TODO: change to reporter.error and add various possible causes in description.
-  // Most likely cause is trying to register a plain object that does not have a
-  // register method and is not a class constructor
-  __DEV__
-    ? createError(`AUR0006: Unable to autoregister dependency: [${deps.map(toStringSafe)}]`)
-    : createError(`AUR0006:${deps.map(toStringSafe)}`);
-const resourceExistError = (key: Key) =>
-  __DEV__
-    ? createError(`AUR0007: Resource key "${toStringSafe(key)}" already registered`)
-    : createError(`AUR0007:${toStringSafe(key)}`);
-const cantResolveKeyError = (key: Key) =>
-  __DEV__
-    ? createError(`AUR0008: Unable to resolve key: ${toStringSafe(key)}`)
-    : createError(`AUR0008:${toStringSafe(key)}`);
-const jitRegisterNonFunctionError = (keyAsValue: Key) =>
-  __DEV__
-    ? createError(`AUR0009: Attempted to jitRegister something that is not a constructor: '${toStringSafe(keyAsValue)}'. Did you forget to register this resource?`)
-    : createError(`AUR0009:${toStringSafe(keyAsValue)}`);
-const jitInstrinsicTypeError = (keyAsValue: any) =>
-  __DEV__
-    ? createError(`AUR0010: Attempted to jitRegister an intrinsic type: ${keyAsValue.name}. Did you forget to add @inject(Key)`)
-    : createError(`AUR0010:${keyAsValue.name}`);
-const invalidResolverFromRegisterError = () =>
-  __DEV__
-    ? createError(`AUR0011: Invalid resolver returned from the static register method`)
-    : createError(`AUR0011`);
-const jitInterfaceError = (name: string) =>
-  __DEV__
-    ? createError(`AUR0012: Attempted to jitRegister an interface: ${name}`)
-    : createError(`AUR0012:${name}`);
 /**
  * An implementation of IRegistry that delegates registration to a
  * separately registered class. The ParameterizedRegistry facilitates the
@@ -1394,24 +880,6 @@ export class ParameterizedRegistry implements IRegistry {
   }
 }
 
-type ResolverLookup = WeakMap<IResolver, unknown>;
-const containerLookup = new WeakMap<IContainer, ResolverLookup>();
-
-function cacheCallbackResult<T>(fun: ResolveCallback<T>): ResolveCallback<T> {
-  return function (handler: IContainer, requestor: IContainer, resolver: IResolver): T {
-    let resolverLookup = containerLookup.get(handler);
-    if (resolverLookup === void 0) {
-      containerLookup.set(handler, resolverLookup = new WeakMap());
-    }
-    if (resolverLookup.has(resolver)) {
-      return resolverLookup.get(resolver) as T;
-    }
-    const t = fun(handler, requestor, resolver);
-    resolverLookup.set(resolver, t);
-    return t;
-  };
-}
-
 /**
  * you can use the resulting {@linkcode IRegistration} of any of the factory methods
  * to register with the container, e.g.
@@ -1422,7 +890,7 @@ function cacheCallbackResult<T>(fun: ResolveCallback<T>): ResolveCallback<T> {
  * container.get(Foo);
  * ```
  */
-export const Registration = {
+ export const Registration = {
   /**
    * allows you to pass an instance.
    * Every time you request this {@linkcode Key} you will get this instance back.
@@ -1433,9 +901,7 @@ export const Registration = {
    * @param key - key to register the instance with
    * @param value - the instance associated with the key
    */
-  instance<T>(key: Key, value: T): IRegistration<T> {
-    return new Resolver(key, ResolverStrategy.instance, value);
-  },
+  instance: instanceRegistration,
   /**
    * Creates an instance from the class.
    * Every time you request this {@linkcode Key} you will get the same one back.
@@ -1446,9 +912,7 @@ export const Registration = {
    * @param key - key to register the singleton class with
    * @param value - the singleton class to instantiate when a container resolves the associated key
    */
-  singleton<T extends Constructable>(key: Key, value: T): IRegistration<InstanceType<T>> {
-    return new Resolver(key, ResolverStrategy.singleton, value);
-  },
+  singleton: singletonRegistration,
   /**
    * Creates an instance from a class.
    * Every time you request this {@linkcode Key} you will get a new instance.
@@ -1459,9 +923,7 @@ export const Registration = {
    * @param key - key to register the transient class with
    * @param value - the class to instantiate when a container resolves the associated key
    */
-  transient<T extends Constructable>(key: Key, value: T): IRegistration<InstanceType<T>> {
-    return new Resolver(key, ResolverStrategy.transient, value);
-  },
+  transient: transientRegistation,
   /**
    * Creates an instance from the method passed.
    * Every time you request this {@linkcode Key} you will get a new instance.
@@ -1473,9 +935,7 @@ export const Registration = {
    * @param key - key to register the callback with
    * @param callback - the callback to invoke when a container resolves the associated key
    */
-  callback<T>(key: Key, callback: ResolveCallback<T>): IRegistration<Resolved<T>> {
-    return new Resolver(key, ResolverStrategy.callback, callback);
-  },
+  callback: callbackRegistration,
   /**
    * Creates an instance from the method passed.
    * On the first request for the {@linkcode Key} your callback is called and returns an instance.
@@ -1490,9 +950,7 @@ export const Registration = {
    * @param key - key to register the cached callback with
    * @param callback - the cache callback to invoke when a container resolves the associated key
    */
-  cachedCallback<T>(key: Key, callback: ResolveCallback<T>): IRegistration<Resolved<T>> {
-    return new Resolver(key, ResolverStrategy.callback, cacheCallbackResult(callback));
-  },
+  cachedCallback: cachedCallbackRegistration,
   /**
    * creates an alternate {@linkcode Key} to retrieve an instance by.
    * Returns the same scope as the original {@linkcode Key}.
@@ -1506,17 +964,13 @@ export const Registration = {
    * @param originalKey - the real key to resolve the get call from a container
    * @param aliasKey - the key that a container allows to resolve the real key associated
    */
-  aliasTo<T>(originalKey: T, aliasKey: Key): IRegistration<Resolved<T>> {
-    return new Resolver(aliasKey, ResolverStrategy.alias, originalKey);
-  },
+  aliasTo: aliasToRegistration,
   /**
    * @internal
    * @param key - the key to register a defer registration
    * @param params - the parameters that should be passed to the resolution of the key
    */
-  defer(key: Key, ...params: unknown[]): IRegistry {
-    return new ParameterizedRegistry(key, params);
-  }
+  defer: deferRegistration,
 };
 
 export class InstanceProvider<K extends Key> implements IDisposableResolver<K | null> {
@@ -1559,43 +1013,10 @@ export class InstanceProvider<K extends Key> implements IDisposableResolver<K | 
   }
 }
 
-function validateKey(key: any): void {
-  if (key === null || key === void 0) {
-    if (__DEV__) {
-      throw createError(`AUR0014: key/value cannot be null or undefined. Are you trying to inject/register something that doesn't exist with DI?`);
-    } else {
-      throw createError(`AUR0014`);
-    }
-  }
-}
-
-function buildAllResponse(resolver: IResolver, handler: IContainer, requestor: IContainer): any[] {
-  if (resolver instanceof Resolver && resolver._strategy === ResolverStrategy.array) {
-    const state = resolver._state as IResolver[];
-    let i = state.length;
-    const results = new Array(i);
-
-    while (i--) {
-      results[i] = state[i].resolve(handler, requestor);
-    }
-
-    return results;
-  }
-
-  return [resolver.resolve(handler, requestor)];
-}
-
 const noInstanceError = (name?: string) => {
   if (__DEV__) {
     return createError(`AUR0013: Cannot call resolve ${name} before calling prepare or after calling dispose.`);
   } else {
     return createError(`AUR0013:${name}`);
   }
-};
-
-const createNativeInvocationError = (Type: Constructable): Error => {
-  if (__DEV__) {
-    return createError(`AUR0015: ${Type.name} is a native function and therefore cannot be safely constructed by DI. If this is intentional, please use a callback or cachedCallback resolver.`);
-  }
-  return createError(`AUR0015:${Type.name}`);
 };
