+++ packages/kernel/src/eventaggregator.ts
@@ -1,4 +1,4 @@
-import { DI } from './di';
+import { createInterface } from './di';
 import { Constructable, IDisposable } from './interfaces';
 import { createError, isString } from './utilities';
 
@@ -7,18 +7,18 @@ import { createError, isString } from './utilities';
  */
 class Handler<T extends Constructable> {
   public constructor(
-    public readonly messageType: T,
-    public readonly callback: (message: InstanceType<T>) => void,
+    private readonly _messageType: T,
+    private readonly _callback: (message: InstanceType<T>) => void,
   ) {}
 
   public handle(message: InstanceType<T>): void {
-    if (message instanceof this.messageType) {
-      this.callback.call(null, message);
+    if (message instanceof this._messageType) {
+      this._callback.call(null, message);
     }
   }
 }
 
-export const IEventAggregator = DI.createInterface<IEventAggregator>('IEventAggregator', x => x.singleton(EventAggregator));
+export const IEventAggregator = createInterface<IEventAggregator>('IEventAggregator', x => x.singleton(EventAggregator));
 export interface IEventAggregator extends EventAggregator {}
 
 /**
@@ -157,7 +157,7 @@ export class EventAggregator {
     channelOrType: string | Constructable,
     callback: (...args: unknown[]) => void,
   ): IDisposable {
-    const sub = this.subscribe(channelOrType as string, function (message, event) {
+    const sub = this.subscribe(channelOrType as string, (message, event) => {
       sub.dispose();
       callback(message, event);
     });
