+++ packages/kernel/src/platform.ts
@@ -1,5 +1,5 @@
 import { Platform } from '@aurelia/platform';
-import { DI } from './di';
+import { createInterface } from './di';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const emptyArray: any[] = Object.freeze<any>([]) as any;
@@ -9,4 +9,4 @@ export const emptyObject: any = Object.freeze({}) as any;
 export function noop(): void {}
 
 export interface IPlatform extends Platform {}
-export const IPlatform = DI.createInterface<IPlatform>('IPlatform');
+export const IPlatform = createInterface<IPlatform>('IPlatform');
